requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
mysql-connector-python==9.0.0
bcrypt==4.1.2
gunicorn==23.0.0
